COM-2294 mutualize things in mixin
