+++ src/core/mixins/learnerDirectoryMixin.js
@@ -1,10 +1,22 @@
+import escapeRegExp from 'lodash/escapeRegExp';
 import { formatIdentity, sortStrings, removeDiacritics } from '@helpers/utils';
 import Users from '@api/Users';
+import Email from '@api/Email';
 import { dateDiff, formatDateDiff } from '@helpers/date';
+import { TRAINEE, DEFAULT_AVATAR } from '@data/constants';
+import { NotifyPositive, NotifyNegative } from '@components/popup/notify';
+import { frPhoneNumber } from '@helpers/vuelidateCustomVal';
+import { required, email } from 'vuelidate/lib/validators';
 
 export const learnerDirectoryMixin = {
   data () {
     return {
+      firstStep: true,
+      loading: false,
+      searchStr: '',
+      learnerCreationModal: false,
+      learnerCreationModalLoading: false,
+      newLearner: { identity: { firstname: '', lastname: '' }, contact: { phone: '' }, local: { email: '' } },
       tableLoading: false,
       learnerList: [],
       pagination: { sortBy: 'name', descending: false, page: 1, rowsPerPage: 15 },
@@ -51,7 +63,31 @@ export const learnerDirectoryMixin = {
       ],
     };
   },
+  validations () {
+    return {
+      newLearner: {
+        identity: { lastname: { required } },
+        local: { email: { required, email } },
+        contact: { phone: { required, frPhoneNumber } },
+      },
+    };
+  },
+  computed: {
+    filteredLearners () {
+      const formattedString = escapeRegExp(removeDiacritics(this.searchStr));
+      return this.learnerList.filter(user => user.learner.noDiacriticsName.match(new RegExp(formattedString, 'i')));
+    },
+  },
   methods: {
+    goToLearnerProfile (row) {
+      this.$router.push({ name: 'ni users learners info', params: { learnerId: row.learner._id } });
+    },
+    updateSearch (value) {
+      this.searchStr = value;
+    },
+    getAvatar (link) {
+      return link || DEFAULT_AVATAR;
+    },
     getDaysSinceLastActivityHistory (lastActivityHistory) {
       if (!lastActivityHistory) return null;
 
@@ -85,6 +121,15 @@ export const learnerDirectoryMixin = {
         this.learnerList = [];
       } finally {
         this.tableLoading = false;
+      }
+    },
+    async sendWelcome () {
+      try {
+        await Email.sendWelcome({ email: this.newLearner.local.email, type: TRAINEE });
+        NotifyPositive('Email envoyé');
+      } catch (e) {
+        console.error(e);
+        NotifyNegative('Erreur lors de l\'envoi du mail.');
       }
     },
   },
